convert into Generic API

+++ MVVMUIKit/Screens/Product/Model/Product.swift
@@ -10,7 +10,7 @@ import Foundation
 struct Product: Decodable {
     let  id: Int
     let  title: String
-    let  price: Float
+    let  price: Double
     let  description: String
     let  category: String
     let  image: String
@@ -19,6 +19,6 @@ struct Product: Decodable {
 }
 
 struct Rate: Decodable {
-    let rate: Float
+    let rate: Double
     let count: Int
 }
